db: enable smart drop/move of temp tables

+++ db/nsr_db_cleanup_only.php
@@ -12,33 +12,47 @@ error_reporting(E_ALL & ~E_NOTICE);
 ////////////////////////////////////////////////////////
 
 include "global_params.inc";
-
 include_once $config_file;
+$db_staging = $DB . "_staging"; 
 
-$db_staging = $DB . "_staging";
+////////////////////////////////////////////////////
+// Confirm operation 
+////////////////////////////////////////////////////
+
+$replace_db_staging = $DB_STAGING_REPLACE ? "TRUE" : "FALSE";
 
 $msg_proceed="
 Clean up Native Species Resolver (NSR) database with the following settings:
   Host: $HOSTNAME
   Database: $DB
   Drop temp tables or move to staging: $DROP_OR_MOVE
-  Staging database name (if applicable): $db_staging
+  " . ($DROP_OR_MOVE=='drop'?"":"Staging database: $db_staging
+  Replace staging: $replace_db_staging") . "
 Enter 'Yes' to proceed, or 'No' to cancel: ";
 $proceed=responseYesNoDie($msg_proceed);
 if ($proceed===false) die("\r\nOperation cancelled\r\n");
+if ( !db_exists($DB, $USER, $PWD) ) die ("\r\nERROR: '$DB': no such database\r\n");
 
-if ( db_exists($DB, $USER, $PWD) ) {
-	// Confirm replacement of entire database if requested
-	if ( $DROP_OR_MOVE == "move" ) {
-		if ( db_exists($db_staging, $USER, $PWD) ) {
-			$msg_conf_replace_db="\r\nPrevious staging database `$db_staging` will be deleted! Are you sure you want to proceed? (Y/N): ";
-			$REPLACE_DB=responseYesNoDie($msg_conf_replace_db);
-			if ($REPLACE_DB===false) die ("\r\nOperation cancelled\r\n");
-		}
+// Check for contradictory parameters and warn if replacing existing database
+if ( $DROP_OR_MOVE == "move" ) {
+	$db_exists=db_exists($db_staging, $USER, $PWD);		
+	$create_db_staging=TRUE;
+	
+	if ( $DB_STAGING_REPLACE==FALSE && $db_exists ) {
+		$create_db_staging=FALSE;	
+	}	
+
+	// Warn if existing database will be deleted
+	if ( $db_exists && $create_db_staging==TRUE ) {
+		$msg_conf_replace_db="\r\nExisting staging database `$db_staging` will be replaced! Are you sure you want to do this? (Y/N): ";
+		$REPLACE_DB=responseYesNoDie($msg_conf_replace_db);
+		if ($REPLACE_DB===false) die ("\r\nOperation cancelled\r\n");
 	}
-} else {
-	die("ERROR: '$DB': no such database!\n");
 }
+
+////////////////////////////////////////////////////
+// Move or drop the tables
+////////////////////////////////////////////////////
 
 // Start timer and connect to mysql
 echo "\r\n********* Begin operation ********* \r\n";
@@ -47,21 +61,22 @@ $echo_on = true;		// Display messages and SQL for debugging
 $SQL_display = true;	// Displays final SQL statement
 
 if ( $DROP_OR_MOVE == "move" ) {
-	echo "Creating staging database:\r\n\r\n";	
-	include "mysql_connect.inc";	// Connect without specifying database
 	
-	// Drop and replace entire database
-	echo "Creating database `$db_staging`...";
-	$sql_create_db="
-		DROP DATABASE IF EXISTS `".$db_staging."`;
-		CREATE DATABASE `".$db_staging."`;
-		USE `".$db_staging."`;
-	";
-	sql_execute_multiple($dbm, $sql_create_db);
-	echo "done\r\n";
-	mysqli_close($dbm);
+	if ( $create_db_staging==TRUE ) {
+		include "mysql_connect.inc";	// Connect without specifying database
+		
+		// Drop and replace entire database
+		echo "Creating database `$db_staging`...";
+		$sql_create_db="
+			DROP DATABASE IF EXISTS `".$db_staging."`;
+			CREATE DATABASE `".$db_staging."`;
+			USE `".$db_staging."`;
+		";
+		sql_execute_multiple($dbm, $sql_create_db);
+		echo "done\r\n";
+		mysqli_close($dbm);
+	}
 	
-	// Replace core tables
 	include "db_connect.inc"; // Reconnect to the new DB 
 	include_once "cleanup_move.inc";
 } else if ( $DROP_OR_MOVE == "drop" ) {
@@ -73,9 +88,9 @@ if ( $DROP_OR_MOVE == "move" ) {
 	die("ERROR: Unknown option '$DROP_OR_MOVE' for parameter \$DROP_OR_MOVE! \n");
 }
 
-//////////////////////////////////////////////////////////////////
+////////////////////////////////////////////////////
 // Close connection and report total time elapsed 
-//////////////////////////////////////////////////////////////////
+////////////////////////////////////////////////////
 
 mysqli_close($dbh);
 include $timer_off;
